Move Vector Table code out of multicore
